back/requirements.txt
django>=5.2.0
django-cors-headers>=4.3.1
openai-whisper>=20231117
ffmpeg-python>=0.2.0
requests>=2.32.3
yt-dlp>=2025.3.0